requirements.txt
actionlib==1.11.13
angles==1.9.11
beautifulsoup4==4.7.1
bondpy==1.8.3
boto3==1.9.188
botocore==1.12.188
bs4==0.0.1
cachetools==3.1.1
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
catkin==0.7.17
certifi==2019.6.16
chardet==3.0.4
Click==7.0
controller-manager==0.15.1
controller-manager-msgs==0.15.1
cv-bridge==1.13.0
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
docutils==0.14
dynamic-reconfigure==1.6.0
fake-useragent==0.1.11
future==0.17.1
gazebo-plugins==2.8.4
gazebo-ros==2.8.4
gencpp==0.6.2
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.12
gennodejs==2.0.1
genpy==0.6.8
gobject==0.1.0
google==2.0.2
google-api-core==1.13.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-vision==0.38.0
googleapis-common-protos==1.6.0
grpcio==1.22.0
gTTS==2.0.3
gTTS-token==1.1.3
httplib2==0.13.0
idna==2.8
image-geometry==1.13.0
interactive-markers==1.11.4
jmespath==0.9.4
kdl-parser-py==1.13.1
laser-geometry==1.6.4
message-filters==1.14.3
numpy==1.16.4
opencv-python==4.1.0.25
playsound==1.2.2
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyAudio==0.2.11
pycairo==1.18.1
PyGObject==3.32.2
PyQt5==5.12.2
PyQt5-sip==4.19.17
pyqtgraph==0.10.0
pyserial==3.4
python-dateutil==2.8.0
python-qt-binding==0.3.5
pytz==2019.1
qt-dotgraph==0.3.11
qt-gui==0.3.11
qt-gui-cpp==0.3.11
qt-gui-py-common==0.3.11
requests==2.22.0
resource-retriever==1.12.4
rosbag==1.14.3
rosboost-cfg==1.14.6
rosclean==1.14.6
roscreate==1.14.6
rosgraph==1.14.3
roslaunch==1.14.3
roslib==1.14.6
roslint==0.11.2
roslz4==1.14.3
rosmake==1.14.6
rosmaster==1.14.3
rosmsg==1.14.3
rosnode==1.14.3
rosparam==1.14.3
rospy==1.14.3
rosservice==1.14.3
rostest==1.14.3
rostopic==1.14.3
rosunit==1.14.6
roswtf==1.14.3
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.10
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.13
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.9
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.10
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rsa==4.0
ruamel.yaml==0.15.99
rviz==1.13.3
s3transfer==0.2.1
selenium==3.141.0
sensor-msgs==1.12.7
six==1.12.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
soupsieve==1.9.2
SpeechRecognition==3.8.1
tf==1.12.0
tf-conversions==1.12.0
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.3
unicode==2.7
Unidecode==1.1.1
urdfdom-py==0.4.0
uritemplate==3.0.0
urllib3==1.25.3
vext==0.7.3
vext.gi==0.7.0
xacro==1.13.3
